move tcjs fetching into hook
